+++ macos/gazepoint_sdk/Package.swift
@@ -4,7 +4,7 @@ import PackageDescription
 let package = Package(
     name: "gazepoint_sdk",
     platforms: [
-        .macOS("12.0")
+        .macOS("13.0")
     ],
     products: [
         .library(name: "gazepoint-sdk", targets: ["gazepoint_sdk"])